Move setup extensions
